Put the core of lookup table parsing in lookup.rs, have separate aranges.rs, pubnames.rs, and pubtypes.rs with implementation specific code.
